Rename conversions package
